requirements.txt
langchain[google-genai]
retry
requests
pandas
sentence_transformers
datasets
torch
numpy
langchain-huggingface
langchain-core
bs4
chromadb
PyMuPDF
python-docx
python-pptx